Refactor getQueryResult to match import pattern for other helpers
