+++ client/routes/workflow/helpers/index.js
@@ -25,6 +25,7 @@ export { default as getEventFullDetails } from './get-event-full-details';
 export { default as getEventSummary } from './get-event-summary';
 export { default as getHistoryEvents } from './get-history-events';
 export { default as getHistoryTimelineEvents } from './get-history-timeline-events';
+export { default as getQueryResult } from './get-query-result';
 export { default as getSummary } from './get-summary';
 export { default as getSummaryWorkflowStatus } from './get-summary-workflow-status';
 export { default as getTimeElapsedDisplay } from './get-time-elapsed-display';
